+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * fix: Fix compute_ids_high_low hint constant path [#2285](https://github.com/lambdaclass/cairo-vm/pull/2285)
 
+* fix: also mark PC as accessed in run_instruction [#2106](https://github.com/lambdaclass/cairo-vm/pull/2106)
+
 * feat: Make Blake2s API public [#2286](https://github.com/lambdaclass/cairo-vm/pull/2286)
 
 * feat: implemented delete_unaccessed function [#2282](https://github.com/lambdaclass/cairo-vm/pull/2282)